Fixed Suggestions
Changed the language tag to typescript and switched the types to typescript.

Removed the unnecessary double binding of the active variable

Removed the .get() functions and replaced them with $'s

And added translator mode settings, key-bind's and the window size to the presets as well as slightly changed the formatting of the code.